// src/models/playlists.ts
// Generated by https://quicktype.io

export interface Playlists {
    name:        string;
    description: string;
    isRanked:    boolean;
    imageUrl:    null | string;
    gameMode:    GameMode;
    isActive:    boolean;
    id:          string;
    contentId:   string;
}

export enum GameMode {
    Arena = "Arena",
    Warzone = "Warzone",
}